requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
oauthlib>=3.1.1
pyOpenSSL>=21.0.0
Flask-Login>=0.5.0